backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# AI & Services
google-generativeai==0.3.2
resend==0.8.0

# Environment & Config
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic[email]

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4